Fixed broken make dependencies
- fixed broken dependencies
- changed type of some instrumented function
- exit code in _rt_end
- reorganized code in rt/

+++ src/executor.cc
@@ -228,7 +228,7 @@ public:
 
       llvm::outs() << "stid: " << i << "\n";
 
-      b.CreateCall (ret, b.getInt32 (retid++));
+      b.CreateCall (ret, b.getInt16 (retid++));
       count++;
    }
